+++ changelog.md
@@ -6,3 +6,5 @@
 - Copied csv files into [EmployeeSQL](EmployeeSQL) folder
 - Created [readme.md](readme.md) and [changelog.md](changelog.md)
 - Created ERD on [QuickDBD](https://app.quickdatabasediagrams.com/#/d/8M7zq1) and saved as [PDF](resources/erd.pdf) and [PNG](resources/erd.png)
+- Created all tables [table_creation.sql](queries/table_creation.sql)
+    - [] Fix data type on ERD departments table